+++ package.json
@@ -13,9 +13,10 @@
     "@testing-library/user-event": "^12.1.7",
     "@types/gapi": "0.0.39",
     "@types/gapi.auth2": "0.0.52",
-    "@types/jest": "^26.0.14",
+    "@types/jest": "^26.0.15",
     "@types/jquery": "^3.3.38",
     "@types/md5": "^2.2.0",
+    "@types/mocha": "^8.0.4",
     "@types/node": "^14.11.8",
     "@types/qrcode.react": "^1.0.0",
     "@types/react": "^16.9.52",
@@ -23,6 +24,7 @@
     "@types/react-dom": "^16.9.7",
     "@types/react-motion": "0.0.29",
     "@types/react-router-dom": "^5.1.5",
+    "@types/react-test-renderer": "^17.0.0",
     "firebase": "^7.23.0",
     "globalize": "^1.6.0",
     "googleapis": "^65.0.0",
@@ -37,6 +39,7 @@
     "react-router": "^5.1.2",
     "react-router-dom": "^5.1.2",
     "react-scripts": "^3.4.4",
+    "react-test-renderer": "^17.0.1",
     "ts-node-dev": "^1.0.0-pre.44",
     "typescript": "^4.0.3",
     "universal-cookie": "^4.0.4"
