+++ src/translations/LanguageContainer.tsx
@@ -3,8 +3,7 @@ import * as React from 'react';
 import messages_en from './locales/en.json';
 import messages_de from './locales/de.json';
 import { IntlProvider } from 'react-intl';
-import { withCookies, Cookies } from 'react-cookie';
-import Routed from '../Routed';
+import Cookies from 'universal-cookie';
 
 interface State {
     locale: string
@@ -16,14 +15,18 @@ const MESSAGES = {
 };
 
 interface Props {
-    cookies: Cookies
 }
 
 class LanguageContainer extends React.Component<Props, State> {
 
+    cookies: Cookies;
+
     constructor(props: Props) {
         super(props);
-        let lang = props.cookies.get("locale");
+
+        this.cookies = new Cookies();
+        
+        let lang = this.cookies.get("locale");
         if (!lang || typeof lang === "undefined") {
             lang = navigator.language.split("-")[0];
         }
@@ -33,17 +36,22 @@ class LanguageContainer extends React.Component<Props, State> {
         this.state = {
             locale: lang
         }
+        this.changeLanguage = this.changeLanguage.bind(this);
+        this.getCurrentLocale = this.getCurrentLocale.bind(this);
     }
 
-    public changeLanguage = (locale: string) => {
-        const { cookies } = this.props;
-        cookies.set("locale", locale, { expires: new Date(9999, 12) })
+    public changeLanguage(locale: string) {
+        this.cookies.set("locale", locale, { expires: new Date(9999, 12) })
         this.setState({
             locale
         });
     }
 
-    hasKey<O>(obj: O, key: keyof any): key is keyof O {
+    public getCurrentLocale(): string {
+        return this.state.locale;
+    }
+
+    private hasKey<O>(obj: O, key: keyof any): key is keyof O {
         return key in obj
     }
 
@@ -56,10 +64,10 @@ class LanguageContainer extends React.Component<Props, State> {
 
         return (
             <IntlProvider locale={locale} messages={msg}>
-                <Routed changeLanguage={this.changeLanguage} currentLocale={locale} />
+                {this.props.children}
             </IntlProvider>
         );
     }
 }
 
-export default withCookies(LanguageContainer);
+export default LanguageContainer;
