+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "target": "es6",
-    "types": ["gapi", "gapi.auth2"],
+    "types": ["gapi", "gapi.auth2", "jest"],
     "lib": [
       "dom",
       "dom.iterable",
